Move extensions to pkg.
